requirements.txt
# API Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
supabase==2.0.2

# AI & Vectorization
openai==1.3.7
anthropic==0.7.7
google-generativeai==0.3.2
groq==0.4.1
cohere==4.37
pinecone-client==2.2.4
sentence-transformers==2.2.2

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0
python-multipart==0.0.6

# Export Libraries (NUEVAS)
weasyprint==61.2
reportlab==4.0.7
markdown==3.5.1

# Utilities
pydantic>=2.7.0
pydantic-settings==2.10.1
python-dotenv==1.0.0
aiofiles==23.2.1

# Database migrations
alembic==1.13.1

# Development
pytest==7.4.3
pytest-asyncio==0.21.1